加入按键调整最大电压功能

+++ board/stm32f103c8t6/src/drv/led.h
@@ -6,7 +6,7 @@
 typedef struct led_s
 {
 	uint8_t led_num;
-	uint8_t led_val;
+	uint16_t led_val;
 } led_s;
 
 void led_init();
